Добавляет модульность в проект
